@@ -9,10 +9,10 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.0"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0"
   },
   "eslintConfig": {
     "extends": "react-app"
