@@ -1,28 +1,24 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
+import React from "react";
+import { useParams, Link } from 'react-router-dom';
 
-class Dog extends Component {
-  render() {
-    let dogFacts = null;
-    if (this.props.currentDog === this.props.name) {
-      dogFacts = (
+function Dog(props) {
+  const params = useParams();
+  const isCurrentDog = params.name === props.name;
+  return (
+    <div>
+      <Link to={`/dogs/${props.name}`}>
+        <h1>{props.name}</h1>
+      </Link>
+      <img src={props.image} alt="dog" />
+      {isCurrentDog &&
         <div>
-          <h1>{`Name: ${this.props.name}`}</h1>
-          <h2>{`Age: ${this.props.age}`}</h2>
-          <ul>{this.props.facts.map((fact, i) => <li key={i}>{fact}</li>)}</ul>
+          <h1>{`Name: ${props.name}`}</h1>
+          <h2>{`Age: ${props.age}`}</h2>
+          <ul>{props.facts.map((fact, i) => <li key={i}>{fact}</li>)}</ul>
         </div>
-      );
-    }
-    return (
-      <div>
-        <Link to={`/dogs/${this.props.name}`}>
-          <h1>{this.props.name}</h1>
-        </Link>
-        <img src={this.props.image} alt="dog picture" />
-        {dogFacts}
-      </div>
-    );
-  }
+      }
+    </div>
+  );
 }
 
 export default Dog;
